+++ extras/Projucer/Source/CodeEditor/jucer_SourceCodeEditor.cpp
@@ -403,7 +403,7 @@ public:
     void paint (Graphics& g) override
     {
         Path outline;
-        outline.addRoundedRectangle (1.0f, 1.0f, getWidth() - 2.0f, getHeight() - 2.0f, 8.0f);
+        outline.addRoundedRectangle (1.0f, 1.0f, (float) getWidth() - 2.0f, (float) getHeight() - 2.0f, 8.0f);
 
         g.setColour (Colours::black.withAlpha (0.6f));
         g.fillPath (outline);
