+++ extras/Projucer/Source/ComponentEditor/Properties/jucer_ColourPropertyComponent.h
@@ -66,7 +66,7 @@ public:
                                 Colour (0xffffffff).overlaidWith (colour));
 
             g.setColour (Colours::white.overlaidWith (colour).contrasting());
-            g.setFont (Font (getHeight() * 0.6f, Font::bold));
+            g.setFont (Font ((float) getHeight() * 0.6f, Font::bold));
             g.drawFittedText (colour.toDisplayString (true),
                               2, 1, getWidth() - 4, getHeight() - 1,
                               Justification::centred, 1);
